% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Function.R
\name{GetPermutePeak}
\alias{GetPermutePeak}
\title{Get random permutation peaks for null distribution}
\usage{
GetPermutePeak(peak.access, permute_peak)
}
\arguments{
\item{peak.access}{Vector of peak names to permute}

\item{permute_peak}{List of pre-computed permutation peaks}
}
\value{
Vector of randomly selected background peaks
}
\description{
Get random permutation peaks for null distribution
}
\keyword{internal}
