% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Function.R
\name{LinkSpatialPeaks}
\alias{LinkSpatialPeaks}
\title{Core function to link peaks to genes}
\usage{
LinkSpatialPeaks(object, res, n.permute = 1000)
}
\arguments{
\item{object}{Seurat object}

\item{res}{Preprocessing results list}

\item{n.permute}{Number of permutations}

\item{cor_method}{Correlation method (default: qlcMatrix::corSparse)}
}
\value{
Data frame with gene-level and peak-level results
}
\description{
Performs gene-level constrained regression and peak-level correlation analysis.
}
\keyword{internal}
