% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Function.R
\name{SpatialPreprocess}
\alias{SpatialPreprocess}
\title{Preprocess spatial data by removing spatial patterns for all features}
\usage{
SpatialPreprocess(expression.data, location, num.core = 1)
}
\arguments{
\item{expression.data}{Matrix of expression or accessibility data (features x spots)}

\item{num.core}{Number of cores for parallel processing (default: 1)}

\item{locations}{Data frame with spatial coordinates}
}
\value{
Matrix of spatially decorrelated data with same dimensions as input
}
\description{
Applies RemoveSpatialPattern to all genes or peaks in parallel.
}
\keyword{internal}
