% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPINK.R
\name{spink_analysis}
\alias{spink_analysis}
\title{Perform SPINK analysis to identify enhancer-gene regulatory links}
\usage{
spink_analysis(object, gene.thr = 0.1, n.permute = 1000)
}
\arguments{
\item{object}{A preprocessed Seurat object from spink_preprocess}

\item{n.permute}{Number of permutations for empirical p-value (default: 1000)}

\item{link.gene.thr}{P-value threshold for gene-level significance (default: 0.1)}
}
\value{
Seurat object with results stored in misc$spink$Link_domain
}
\description{
This function performs gene-level and peak-level statistical testing
to identify significant enhancer-gene regulatory interactions.
}
